requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
coreschema==0.0.4
Django==4.0.1
django-autocomplete-light==3.8.2
django-jalali-date==1.0.1
djangorestframework==3.13.1
idna==3.3
itypes==1.2.0
jdatetime==3.8.1
Jinja2==3.0.3
MarkupSafe==2.0.1
Pillow==9.0.0
psycopg2-binary==2.9.3
Pygments==2.11.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.8
